refactor to use separate process for image manipulation
